Moved generic auto_blacklist function from uds module to can_actions
